simplify submit function
